agregando test a ms de productos

+++ domain/ordermgmt-product/src/main/java/com/ordermgmt/products/repository/IProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IProductRepository extends JpaRepository<Product, Integer> {
+public interface IProductRepository extends JpaRepository<Product, Long> {
 
     List<Product> findByIdIn(List<Long> ids);
     boolean existsByProductName(String productName);
